+++ packages/oidc/src/get-vercel-oidc-token.ts
@@ -1,4 +1,5 @@
 import { getContext } from './get-context';
+
 /**
  * Gets the current OIDC token from the request context or the environment variable.
  *
@@ -6,6 +7,8 @@ import { getContext } from './get-context';
  *
  * This function is used to retrieve the OIDC token from the request context or the environment variable.
  * It checks for the `x-vercel-oidc-token` header in the request context and falls back to the `VERCEL_OIDC_TOKEN` environment variable if the header is not present.
+ *
+ * Unlike the `getVercelOidcTokenSync` function, this function will refresh the token if it is expired in a development environment.
  *
  * @returns {Promise<string>} A promise that resolves to the OIDC token.
  * @throws {Error} If the `x-vercel-oidc-token` header is missing from the request context and the environment variable `VERCEL_OIDC_TOKEN` is not set.
@@ -22,7 +25,29 @@ import { getContext } from './get-context';
  * ```
  */
 export async function getVercelOidcToken(): Promise<string> {
-  return getVercelOidcTokenSync();
+  let token = '';
+  let err: any;
+
+  try {
+    token = getVercelOidcTokenSync();
+  } catch (error) {
+    err = error;
+  }
+
+  try {
+    const { getTokenPayload, isExpired } = await import('./token-util');
+    const { refreshToken } = await import('./token');
+    if (!token || isExpired(getTokenPayload(token))) {
+      await refreshToken();
+      token = getVercelOidcTokenSync();
+    }
+  } catch (error) {
+    if (error instanceof Error && err.message === error.message) {
+      throw err;
+    }
+    throw new Error(`Failed to refresh OIDC token: ${err}: ${error}`);
+  }
+  return token;
 }
 
 /**
@@ -32,6 +57,8 @@ export async function getVercelOidcToken(): Promise<string> {
  *
  * This function is used to retrieve the OIDC token from the request context or the environment variable.
  * It checks for the `x-vercel-oidc-token` header in the request context and falls back to the `VERCEL_OIDC_TOKEN` environment variable if the header is not present.
+ *
+ * This function will not refresh the token if it is expired. For refreshing the token, use the @{link getVercelOidcToken} function.
  *
  * @returns {string} The OIDC token.
  * @throws {Error} If the `x-vercel-oidc-token` header is missing from the request context and the environment variable `VERCEL_OIDC_TOKEN` is not set.